+++ extra_views/__init__.py
@@ -1,5 +1,9 @@
-from extra_views.formsets import FormSetView, ModelFormSetView, InlineFormSetView
-from extra_views.advanced import CreateWithInlinesView, UpdateWithInlinesView, InlineFormSet, NamedFormsetsMixin
+from extra_views.formsets import FormSetView, ModelFormSetView, \
+    InlineFormSetView, BaseFormSetMixin, BaseInlineFormSetMixin
+from extra_views.advanced import CreateWithInlinesView, \
+    UpdateWithInlinesView, InlineFormSetFactory, NamedFormsetsMixin, InlineFormSet, \
+    FormSetSuccessMessageMixin, SuccessMessageMixin
 from extra_views.dates import CalendarMonthView
 from extra_views.contrib.mixins import SearchableListMixin, SortableListMixin
 
+__version__ = "0.12.0"