+++ extra_views/formsets.py
@@ -1,6 +1,3 @@
-from functools import partial
-from functools import wraps
-
 import django
 from django.views.generic.base import TemplateResponseMixin, View, ContextMixin
 from django.http import HttpResponseRedirect
@@ -8,55 +5,53 @@ from django.forms.formsets import formset_factory
 from django.forms.models import modelformset_factory, inlineformset_factory
 from django.views.generic.detail import SingleObjectMixin, SingleObjectTemplateResponseMixin
 from django.views.generic.list import MultipleObjectMixin, MultipleObjectTemplateResponseMixin
-from django.forms.models import BaseInlineFormSet
 
 
-class BaseFormSetMixin(object):
+class BaseFormSetFactory(object):
     """
-    Base class for constructing a FormSet within a view
+    Base class for constructing a FormSet from `formset_factory` in a view.
+
+    Calling `construct_formset` calls all other methods.
     """
 
     initial = []
     form_class = None
     formset_class = None
-    success_url = None
-    extra = 2
-    max_num = None
-    can_order = False
-    can_delete = False
     prefix = None
+    formset_kwargs = {}
+    factory_kwargs = {}
 
     def construct_formset(self):
         """
         Returns an instance of the formset
         """
         formset_class = self.get_formset()
-        extra_form_kwargs = self.get_extra_form_kwargs()
-
-        # Hack to let as pass additional kwargs to each forms constructor. Be aware that this
-        # doesn't let us provide *different* arguments for each form
-        if extra_form_kwargs:
-            formset_class.form = wraps(formset_class.form)(partial(formset_class.form, **extra_form_kwargs))
-
+        if hasattr(self, 'get_extra_form_kwargs'):
+            klass = type(self).__name__
+            raise DeprecationWarning(
+                'Calling {0}.get_extra_form_kwargs is no longer supported. '
+                'Set `form_kwargs` in {0}.formset_kwargs or override '
+                '{0}.get_formset_kwargs() directly.'.format(klass),
+            )
         return formset_class(**self.get_formset_kwargs())
 
     def get_initial(self):
         """
-        Returns the initial data to use for formsets on this view.
+        Returns a copy of the initial data to use for formsets on this view.
         """
-        return self.initial
+        return self.initial[:]
+
+    def get_prefix(self):
+        """
+        Returns the prefix used for formsets on this view.
+        """
+        return self.prefix
 
     def get_formset_class(self):
         """
         Returns the formset class to use in the formset factory
         """
         return self.formset_class
-
-    def get_extra_form_kwargs(self):
-        """
-        Returns extra keyword arguments to pass to each form in the formset
-        """
-        return {}
 
     def get_form_class(self):
         """
@@ -74,21 +69,15 @@ class BaseFormSetMixin(object):
         """
         Returns the keyword arguments for instantiating the formset.
         """
-        kwargs = {}
-
-        # We have to check whether initial has been set rather than blindly passing it along,
-        # This is because Django 1.3 doesn't let inline formsets accept initial, and no versions
-        # of Django let generic inline formset handle initial data.
-        initial = self.get_initial()
-        if initial:
-            kwargs['initial'] = initial
-
-        if self.prefix:
-            kwargs['prefix'] = self.prefix
+        kwargs = self.formset_kwargs.copy()
+        kwargs.update({
+            'initial': self.get_initial(),
+            'prefix': self.get_prefix(),
+        })
 
         if self.request.method in ('POST', 'PUT'):
             kwargs.update({
-                'data': self.request.POST,
+                'data': self.request.POST.copy(),
                 'files': self.request.FILES,
             })
         return kwargs
@@ -97,23 +86,36 @@ class BaseFormSetMixin(object):
         """
         Returns the keyword arguments for calling the formset factory
         """
-        kwargs = {
-            'extra': self.extra,
-            'max_num': self.max_num,
-            'can_order': self.can_order,
-            'can_delete': self.can_delete
-        }
+        # Perform deprecation check
+        for attr in ['extra', 'max_num', 'can_order', 'can_delete', 'ct_field',
+                     'formfield_callback', 'fk_name', 'widgets', 'ct_fk_field']:
+            if hasattr(self, attr):
+                klass = type(self).__name__
+                raise DeprecationWarning(
+                    'Setting `{0}.{1}` at the class level is now deprecated. '
+                    'Set `{0}.factory_kwargs` instead.'.format(klass, attr)
+                )
 
+        kwargs = self.factory_kwargs.copy()
         if self.get_formset_class():
             kwargs['formset'] = self.get_formset_class()
-
         return kwargs
 
 
-class FormSetMixin(BaseFormSetMixin, ContextMixin):
+class BaseFormSetMixin(BaseFormSetFactory):
+    def __init__(self, *args, **kwargs):
+        from warnings import warn
+        warn('`extra_views.BaseFormSetMixin` has been renamed to '
+             '`BaseFormSetFactory`. `BaseFormSetMixin` will be removed in '
+             'a future release.', DeprecationWarning)
+        super(BaseFormSetMixin, self).__init__(*args, **kwargs)
+
+
+class FormSetMixin(BaseFormSetFactory, ContextMixin):
     """
-    A mixin that provides a way to show and handle a formset in a request.
+    A view mixin that provides a way to show and handle a single formset in a request.
     """
+    success_url = None
 
     def get_success_url(self):
         """
@@ -142,32 +144,13 @@ class FormSetMixin(BaseFormSetMixin, ContextMixin):
 
 class ModelFormSetMixin(FormSetMixin, MultipleObjectMixin):
     """
-    A mixin that provides a way to show and handle a model formset in a request.
+    A view mixin that provides a way to show and handle a single model formset in a request.
+
+    Uses `modelformset_factory`.
     """
 
     exclude = None
     fields = None
-    formfield_callback = None
-    widgets = None
-
-    def get_context_data(self, **kwargs):
-        """
-        If an object list has been supplied, inject it into the context with the
-        supplied context_object_name name.
-        """
-        context = {}
-
-        if self.object_list is not None:
-            context['object_list'] = self.object_list
-            context_object_name = self.get_context_object_name(self.object_list)
-            if context_object_name:
-                context[context_object_name] = self.object_list
-        context.update(kwargs)
-
-        # MultipleObjectMixin get_context_data() doesn't work when object_list
-        # is not provided in kwargs, so we skip MultipleObjectMixin and call
-        # ContextMixin directly.
-        return ContextMixin.get_context_data(self, **context)
 
     def get_formset_kwargs(self):
         """
@@ -182,16 +165,11 @@ class ModelFormSetMixin(FormSetMixin, MultipleObjectMixin):
         Returns the keyword arguments for calling the formset factory
         """
         kwargs = super(ModelFormSetMixin, self).get_factory_kwargs()
-        kwargs.update({
-            'exclude': self.exclude,
-            'fields': self.fields,
-            'formfield_callback': self.formfield_callback,
-            'widgets': self.widgets,
-        })
+        kwargs.setdefault('fields', self.fields)
+        kwargs.setdefault('exclude', self.exclude)
+
         if self.get_form_class():
             kwargs['form'] = self.get_form_class()
-        if self.get_formset_class():
-            kwargs['formset'] = self.get_formset_class()
         return kwargs
 
     def get_formset(self):
@@ -208,34 +186,16 @@ class ModelFormSetMixin(FormSetMixin, MultipleObjectMixin):
         return super(ModelFormSetMixin, self).formset_valid(formset)
 
 
-class BaseInlineFormSetMixin(BaseFormSetMixin):
+class BaseInlineFormSetFactory(BaseFormSetFactory):
     """
-    Base class for constructing an inline formSet within a view
+    Base class for constructing a FormSet from `inlineformset_factory` in a view.
+
+    Calling `construct_formset` calls all other methods.
     """
     model = None
     inline_model = None
-    fk_name = None
-    formset_class = BaseInlineFormSet
     exclude = None
     fields = None
-    formfield_callback = None
-    can_delete = True
-    save_as_new = False
-
-    def get_context_data(self, **kwargs):
-        """
-        If an object has been supplied, inject it into the context with the
-        supplied context_object_name name.
-        """
-        context = {}
-
-        if self.object:
-            context['object'] = self.object
-            context_object_name = self.get_context_object_name(self.object)
-            if context_object_name:
-                context[context_object_name] = self.object
-        context.update(kwargs)
-        return super(BaseInlineFormSetMixin, self).get_context_data(**context)
 
     def get_inline_model(self):
         """
@@ -247,8 +207,14 @@ class BaseInlineFormSetMixin(BaseFormSetMixin):
         """
         Returns the keyword arguments for instantiating the formset.
         """
-        kwargs = super(BaseInlineFormSetMixin, self).get_formset_kwargs()
-        kwargs['save_as_new'] = self.save_as_new
+        # Perform deprecation check
+        if hasattr(self, 'save_as_new'):
+            klass = type(self).__name__
+            raise DeprecationWarning(
+                'Setting `{0}.save_as_new` at the class level is now '
+                'deprecated. Set `{0}.formset_kwargs` instead.'.format(klass)
+            )
+        kwargs = super(BaseInlineFormSetFactory, self).get_formset_kwargs()
         kwargs['instance'] = self.object
         return kwargs
 
@@ -256,17 +222,12 @@ class BaseInlineFormSetMixin(BaseFormSetMixin):
         """
         Returns the keyword arguments for calling the formset factory
         """
-        kwargs = super(BaseInlineFormSetMixin, self).get_factory_kwargs()
-        kwargs.update({
-            'exclude': self.exclude,
-            'fields': self.fields,
-            'formfield_callback': self.formfield_callback,
-            'fk_name': self.fk_name,
-        })
+        kwargs = super(BaseInlineFormSetFactory, self).get_factory_kwargs()
+        kwargs.setdefault('fields', self.fields)
+        kwargs.setdefault('exclude', self.exclude)
+
         if self.get_form_class():
             kwargs['form'] = self.get_form_class()
-        if self.get_formset_class():
-            kwargs['formset'] = self.get_formset_class()
         return kwargs
 
     def get_formset(self):
@@ -276,9 +237,18 @@ class BaseInlineFormSetMixin(BaseFormSetMixin):
         return inlineformset_factory(self.model, self.get_inline_model(), **self.get_factory_kwargs())
 
 
-class InlineFormSetMixin(BaseInlineFormSetMixin, FormSetMixin, SingleObjectMixin):
+class BaseInlineFormSetMixin(BaseInlineFormSetFactory):
+    def __init__(self, *args, **kwargs):
+        from warnings import warn
+        warn('`extra_views.BaseInlineFormSetMixin` has been renamed to '
+             '`BaseInlineFormSetFactory`. `BaseInlineFormSetMixin` will be removed in '
+             'a future release.', DeprecationWarning)
+        super(BaseInlineFormSetMixin, self).__init__(*args, **kwargs)
+
+
+class InlineFormSetMixin(BaseInlineFormSetFactory, SingleObjectMixin, FormSetMixin):
     """
-    A mixin that provides a way to show and handle a inline formset in a request.
+    A view mixin that provides a way to show and handle a single inline formset in a request.
     """
 
     def formset_valid(self, formset):