+++ extra_views_tests/views.py
@@ -1,28 +1,46 @@
-from extra_views import FormSetView, ModelFormSetView, InlineFormSetView, InlineFormSet, CreateWithInlinesView, UpdateWithInlinesView, CalendarMonthView, NamedFormsetsMixin, SortableListMixin, SearchableListMixin
-from extra_views.generic import GenericInlineFormSet, GenericInlineFormSetView
+from extra_views import (
+    FormSetView, ModelFormSetView, InlineFormSetView, InlineFormSetFactory,
+    CreateWithInlinesView, UpdateWithInlinesView, CalendarMonthView,
+    NamedFormsetsMixin, SortableListMixin, SearchableListMixin,
+    SuccessMessageMixin, FormSetSuccessMessageMixin
+)
+from extra_views.generic import GenericInlineFormSetFactory, GenericInlineFormSetView
 from django.views import generic
 from .forms import AddressForm, ItemForm, OrderForm
 from .formsets import BaseArticleFormSet
 from .models import Item, Order, Tag, Event
 
 
-class AddressFormSetView(FormSetView):
+class AddressFormSetView(FormSetSuccessMessageMixin, FormSetView):
     form_class = AddressForm
     template_name = 'extra_views/address_formset.html'
-
-    def get_extra_form_kwargs(self):
-        return {
-            'user': 'foo',
-        }
+    success_message = 'Formset objects were created successfully!'
 
 
 class AddressFormSetViewNamed(NamedFormsetsMixin, AddressFormSetView):
     inlines_names = ['AddressFormset']
 
 
+class AddressFormSetViewKwargs(FormSetView):
+    # Used for testing class level kwargs
+    form_class = AddressForm
+    template_name = 'extra_views/address_formset.html'
+    formset_kwargs = {'auto_id': 'id_test_%s',
+                      'form_kwargs': {'empty_permitted': True}}
+    factory_kwargs = {'max_num': 27}
+    prefix = 'test_prefix'
+    initial = [{'name': 'address1'}]
+
+
 class ItemModelFormSetView(ModelFormSetView):
     model = Item
     fields = ['name', 'sku', 'price', 'order', 'status']
+    template_name = 'extra_views/item_formset.html'
+
+
+class ItemModelFormSetExcludeView(ModelFormSetView):
+    model = Item
+    exclude = ['sku', 'price']
     template_name = 'extra_views/item_formset.html'
 
 
@@ -46,25 +64,23 @@ class PagedModelFormSetView(ModelFormSetView):
     template_name = 'extra_views/paged_formset.html'
 
 
-class ItemsInline(InlineFormSet):
+class ItemsInline(InlineFormSetFactory):
     model = Item
     fields = ['name', 'sku', 'price', 'order', 'status']
 
 
-class TagsInline(GenericInlineFormSet):
+class TagsInline(GenericInlineFormSetFactory):
     model = Tag
     fields = ['name']
 
 
-class OrderCreateView(CreateWithInlinesView):
+class OrderCreateView(SuccessMessageMixin, CreateWithInlinesView):
     model = Order
     fields = ['name']
     context_object_name = 'order'
     inlines = [ItemsInline, TagsInline]
     template_name = 'extra_views/order_and_items.html'
-
-    def get_success_url(self):
-        return '/inlines/%i' % self.object.pk
+    success_message = 'Order %(name)s was created successfully!'
 
     def form_valid(self, form):
         response = super(OrderCreateView, self).form_valid(form)
@@ -83,14 +99,12 @@ class OrderUpdateView(UpdateWithInlinesView):
     inlines = [ItemsInline, TagsInline]
     template_name = 'extra_views/order_and_items.html'
 
-    def get_success_url(self):
-        return ''
-
 
 class OrderTagsView(GenericInlineFormSetView):
     model = Order
     inline_model = Tag
     template_name = "extra_views/inline_formset.html"
+    initial = [{'name': 'test_tag_name'}]
 
 
 class EventCalendarView(CalendarMonthView):